move storage2
