+++ discovery_engine_core/core/src/storage/sqlite/dart_migrations.rs
@@ -24,10 +24,11 @@ use crate::{
     stack,
     storage::{
         models::{NewDocument, NewsResource, NewscatcherData},
+        DartMigrationData,
         Error,
+        MigrationDocument,
         Storage,
     },
-    storage2::{DartMigrationData, MigrationDocument},
 };
 
 use super::SqliteStorage;
